backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Scientific Computing
numpy==1.26.3
scipy==1.12.0

# File Processing
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.8
ezdxf==1.1.4

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
